refactor(qbittorrent-e2e): group scenario steps into fixtures/ and qbittorrent/ subfolders
