+++ src/console/ci/qbittorrent/scenario_steps/fixtures/build_payload_fixture.rs
@@ -1,14 +1,14 @@
-use super::super::torrent_artifacts::build_payload_bytes;
+use super::super::super::torrent_artifacts::build_payload_bytes;
 
 /// In-memory payload fixture used to generate torrent metadata and integrity checks.
-pub(in super::super) struct GeneratedPayload {
-    pub(in super::super) bytes: Vec<u8>,
+pub struct GeneratedPayload {
+    pub bytes: Vec<u8>,
 }
 
 /// Builds deterministic payload bytes for the E2E scenario.
 ///
 /// The generated payload is stable for a given size, which keeps test behavior reproducible.
-pub(in super::super) fn build_payload_fixture(payload_size_bytes: usize) -> GeneratedPayload {
+pub fn build_payload_fixture(payload_size_bytes: usize) -> GeneratedPayload {
     GeneratedPayload {
         bytes: build_payload_bytes(payload_size_bytes),
     }
