+++ src/console/ci/qbittorrent/scenario_steps/fixtures/build_torrent_fixture.rs
@@ -1,11 +1,11 @@
 use anyhow::Context;
 
-use super::super::torrent_artifacts::build_torrent_bytes;
+use super::super::super::torrent_artifacts::build_torrent_bytes;
 use super::build_payload_fixture::GeneratedPayload;
 
 /// In-memory `.torrent` fixture generated from a payload fixture.
-pub(in super::super) struct GeneratedTorrent {
-    pub(in super::super) bytes: Vec<u8>,
+pub struct GeneratedTorrent {
+    pub bytes: Vec<u8>,
 }
 
 /// Builds torrent metadata bytes from a payload fixture.
@@ -13,7 +13,7 @@ pub(in super::super) struct GeneratedTorrent {
 /// # Errors
 ///
 /// Returns an error when torrent metadata encoding fails.
-pub(in super::super) fn build_torrent_fixture(
+pub fn build_torrent_fixture(
     payload: &GeneratedPayload,
     payload_name: &str,
     announce_url: &str,
